requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
keras>=2.15.0
numpy
pandas
tensorflow
tensorflow_intel
yfinance
transformers
scikit_learn
ta
python-decouple
feedparser
newsdataapi
tf-keras
statsmodels>=0.14.2
torch
forex-python